% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_next_integer.R
\name{calculate_next_integer}
\alias{calculate_next_integer}
\title{Calculates maximum of a vector and adds one after removing all letters despite casing}
\usage{
calculate_next_integer(id_vector, sampled = FALSE)
}
\arguments{
\item{id_vector}{vector of length of 1 or greater of combination of letters and strings.}

\item{sampled}{if TRUE, will add a number from 1 to 9 at random instead of adding 1.}
}
\description{
Calculates maximum of a vector and adds one after removing all letters despite casing
}
