% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carry_forward_and_add_one.R
\name{carry_forward_and_add_one}
\alias{carry_forward_and_add_one}
\title{Imputes true NAs with the prior value in a character string (excluding position 1)}
\usage{
carry_forward_and_add_one(vector)
}
\arguments{
\item{vector}{character vector of length 2 or greater}
}
\description{
Imputes true NAs with the prior value in a character string (excluding position 1)
}
