% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluck_words.R
\name{cherrypick_words_2}
\alias{cherrypick_words_2}
\title{Get words following a phrase based on number or}
\usage{
cherrypick_words_2(x, phrase, n_words = 3L, upto = NULL)
}
\description{
Get words following a phrase based on number or
}
