% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_phrase_from_string.R
\name{delete_phrase_from_string}
\alias{delete_phrase_from_string}
\title{Deletes a phrase from a string}
\usage{
delete_phrase_from_string(string, phrase)
}
\arguments{
\item{string}{character vector of length 1}

\item{phrase}{character vector of length 1 of the phrase to be deleted}
}
\description{
Deletes a phrase from a string
}
