% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_whitespace_in_string.R
\name{delete_whitespace_in_string}
\alias{delete_whitespace_in_string}
\title{Delete extra spaces}
\usage{
delete_whitespace_in_string(character_vector)
}
\arguments{
\item{character_vector}{vector of strings of length 1 or greater}
}
\description{
Delete extra spaces
}
