% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_between_vectors.R
\name{diff_between_vectors}
\alias{diff_between_vectors}
\title{Get what is in vector that is not in the other vector}
\usage{
diff_between_vectors(vector, other_vector)
}
\arguments{
\item{vector}{main vector that is being assessed}

\item{other_vector}{vector that is being compared against}
}
\description{
Get what is in vector that is not in the other vector
}
