% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_between_strings.R
\name{diff_strings}
\alias{diff_strings}
\title{Get the difference between a string and a "subtraction" string}
\usage{
diff_strings(string, subtraction_string)
}
\arguments{
\item{string}{main string that is being subtracted from}

\item{subtraction_string}{the string that is being removed from the primary string}
}
\description{
Get the difference between a string and a "subtraction" string
}
