% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erase_phrase.R
\name{erase_phrase}
\alias{erase_phrase}
\title{Removes a pharse from a string}
\usage{
erase_phrase(string, phrase)
}
\arguments{
\item{string}{character vector of length 1}

\item{phrase}{character vector of length 1 to be removed from string}
}
\description{
Removes a pharse from a string
}
