% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_word_by_index.R
\name{extract_word_by_index}
\alias{extract_word_by_index}
\title{Return the word in the string by number}
\usage{
extract_word_by_index(string, n)
}
\arguments{
\item{string}{character vector of length one}

\item{n}{word number}
}
\description{
Return the word in the string by number
}
