% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_leading_decimal.R
\name{has_leading_decimal}
\alias{has_leading_decimal}
\title{Logical for strings that start with 1 or more decimal points}
\usage{
has_leading_decimal(strings)
}
\description{
Logical for strings that start with 1 or more decimal points
}
