% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_leading_pattern.R
\name{has_leading_pattern}
\alias{has_leading_pattern}
\title{Logical Function checking for leading patterns}
\usage{
has_leading_pattern(strings, pattern)
}
\arguments{
\item{pattern}{expression in a non-regex format that is automatically converted to the {pattern}$ in the function}
}
\description{
Logical Function checking for leading patterns
}
