% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_leading_punctuation.R
\name{has_leading_punct}
\alias{has_leading_punct}
\title{Logical for strings that have 1 or more leading punctuation marks}
\usage{
has_leading_punct(strings)
}
\description{
Logical for strings that have 1 or more leading punctuation marks
}
