% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_regex.R
\name{has_regex}
\alias{has_regex}
\title{Logical Function checking for leading patterns}
\usage{
has_regex(strings, regex)
}
\arguments{
\item{regex}{Full regex expression.}
}
\description{
This function is meant to be a generic for any of the logical functions.
}
