% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_trailing_pattern.R
\name{has_trailing_pattern}
\alias{has_trailing_pattern}
\title{Logical Function checking for trailing patterns}
\usage{
has_trailing_pattern(strings, pattern)
}
\arguments{
\item{pattern}{expression in a non-regex format that is automatically converted to the {pattern}$ in the function}
}
\description{
Logical Function checking for trailing patterns
}
