% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_unclosed_curly_brackets.R
\name{has_unclosed_curly_brackets}
\alias{has_unclosed_curly_brackets}
\title{Logical for unclosed curly brackets}
\usage{
has_unclosed_curly_brackets(strings)
}
\description{
Logical for unclosed curly brackets
}
