% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_unclosed_punct.R
\name{has_unclosed_elements}
\alias{has_unclosed_elements}
\title{Aggregates the logic functions for all unclosed elements}
\usage{
has_unclosed_elements(strings)
}
\description{
Aggregate of parentheses, brackets, and curly bracket logical functions.
}
