% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_unclosed_parentheses.R
\name{has_unclosed_parentheses}
\alias{has_unclosed_parentheses}
\title{Logical for unclosed parentheses}
\usage{
has_unclosed_parentheses(strings)
}
\description{
Logical for unclosed parentheses
}
