% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_title_format.R
\name{in_title_format}
\alias{in_title_format}
\title{Format a string with title-style capitalization}
\usage{
in_title_format(string, exclusions = c("of", "the", "in", "and"))
}
\arguments{
\item{string}{string to be formatted}

\item{exclusions}{character vector of exclusions in the expected output format}
}
\description{
Format a string with title-style capitalization
}
