% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_blank.R
\name{no_blank}
\alias{no_blank}
\title{Remove blank from a string}
\usage{
no_blank(vector, all_lengths = FALSE)
}
\arguments{
\item{all_lengths}{if TRUE, strings of one space (ie "^ $") and longer (ie "^          $") are also removed.}
}
\description{
Remove blank from a string
}
