% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_in_words.R
\name{number_in_words}
\alias{number_in_words}
\title{Converts numbers to numbers as words}
\usage{
number_in_words(x)
}
\arguments{
\item{x}{character vector of numbers}
}
\description{
Converts numbers to numbers as words
}
