% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste0_if_not_na.R
\name{paste0_if_not_na}
\alias{paste0_if_not_na}
\title{Paste0 for character_vector into string after removing true NA}
\usage{
paste0_if_not_na(character_vector, ...)
}
\arguments{
\item{character_vector}{vector greater than length 1}
}
\description{
Paste0 for character_vector into string after removing true NA
}
