% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_if_not_na.R
\name{paste_if_not_na}
\alias{paste_if_not_na}
\title{Paste everything but na into a single string}
\usage{
paste_if_not_na(..., sep = "", collapse = " ")
}
\arguments{
\item{...}{character vector to convert to string}
}
\description{
Paste everything but na into a single string
}
