% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluck_words.R
\name{pluck_words}
\alias{pluck_words}
\title{Get words following a phrase based on number or a seocnd phrase}
\usage{
pluck_words(x, phrase, n_words = 3L, upto = NULL)
}
\description{
Get words following a phrase based on number or a seocnd phrase
}
