% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_phrase_from_string.R
\name{remove_phrase_from_string}
\alias{remove_phrase_from_string}
\title{Removes a pharse from a string}
\usage{
remove_phrase_from_string(string, phrase)
}
\arguments{
\item{string}{character vector of length 1}

\item{phrase}{character vector of length 1 to be removed from string}
}
\description{
Removes a pharse from a string
}
