% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_true_na.R
\name{remove_true_na}
\alias{remove_true_na}
\title{Remove true na from a character vector}
\usage{
remove_true_na(character_vector)
}
\arguments{
\item{character_vector}{vector greater than 0}
}
\description{
Remove true na from a character vector
}
