% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_space_before_period.R
\name{rm_space_before_period}
\alias{rm_space_before_period}
\title{Remove spaces before period}
\usage{
rm_space_before_period(x)
}
\description{
This function takes a string such as "Mary had a little lamb ." or "Mary had a little lamb ..." and removes the unecessary space to "Mary had a little lamb.".
}
