% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_random.R
\name{str_random}
\alias{str_random}
\title{Creates a random string of either all uppercase, all lowercase, or both of a given character_length}
\usage{
str_random(character_length, replace = TRUE, type = "both")
}
\description{
Creates a random string of either all uppercase, all lowercase, or both of a given character_length
}
