% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_remove_all_letters.R
\name{str_remove_all_letters}
\alias{str_remove_all_letters}
\title{Removes all letters in a string}
\usage{
str_remove_all_letters(x, type = c("both", "upper", "lower"))
}
\description{
Removes all letters in a string
}
