% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_replace_loop.R
\name{str_replace_loop}
\alias{str_replace_loop}
\title{Loop over replacements in a string}
\usage{
str_replace_loop(x, pattern_vector, replacement_vector)
}
\description{
Loop over replacements in a string
}
