% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit_trim_pipe.R
\name{strsplit_trim_pipe}
\alias{strsplit_trim_pipe}
\title{Splits a string by the pipe, unlists, and trims whitespace on both sides}
\usage{
strsplit_trim_pipe(string)
}
\description{
Splits a string by the pipe, unlists, and trims whitespace on both sides
}
