% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_inner_space.R
\name{trim_inner_space}
\alias{trim_inner_space}
\title{Trims extra spaces within strings}
\usage{
trim_inner_space(vector)
}
\arguments{
\item{vector}{character vector of length one or greater}
}
\value{
vector with inner whitespace removed
}
\description{
Trims extra spaces within strings
}
