% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimws.R
\name{trimws}
\alias{trimws}
\title{Base trimws with an "inner" option that removes all spaces greater than 1 inside strings}
\usage{
trimws(
  x,
  which = c("both", "left", "right"),
  whitespace = "[ \\t\\r\\n]",
  inner_space = TRUE
)
}
\description{
Base trimws with an "inner" option that removes all spaces greater than 1 inside strings
}
