% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_next_integer.R
\name{calculate_next_integer}
\alias{calculate_next_integer}
\title{Calculates maximum of a vector and adds one}
\usage{
calculate_next_integer(vector, sampled = FALSE)
}
\arguments{
\item{vector}{vector of 1 or more integers}

\item{sampled}{if TRUE, will add a number from 1 to 9 at random instead of adding 1.}
}
\description{
Calculates maximum of a vector and adds one
}
